+++ src/gui/checkbox.cxx
@@ -48,7 +48,7 @@ Checkbox::draw(DrawingContext& gc)
 	if (is_checked)
 		gc.draw(checkmark, pos);
 	
-	gc.print_right(Fonts::smallfont, (int)pos.x, (int)pos.y, label);
+	gc.print_right(Fonts::smallfont, pos.x, pos.y, label);
 }
 
 bool
